web interface/requirements.txt
fastapi
pydantic
pymodbus
flask
opcua
python-snap7